Move navigation links to their own file, import where needed, closes #133
Include icons and `iconStyles` in `navLinks`, use `createElement` to render icons in their own component
